adjust header files
make nav bar profile bar into seperate file to see easier
