--- src/components/header-footer/NavCopy.js
@@ -1,132 +1,0 @@
-import React from 'react'
-import { FaBars } from 'react-icons/fa'
-import { NavLink as Link } from 'react-router-dom'
-import styled from 'styled-components'
-
-/**
- * Defines structure and functionality of navbar
- * using the components made in NavbarElements.js
- */
-
-// main navbar container
-const Nav = styled.nav`
-  background: #63d471;
-  border-top: 2px solid #1a3636;
-  border-bottom: 2px solid #1a3636;
-  height: 85px;
-  display: flex;
-  justify-content: space-between;
-  padding: 0.2rem calc((100vw - 1000px) / 2);
-  z-index: 12;
-  /* Third Nav */
-  /* justify-content: flex-start; */
-`
-
-// styled NavLink component
-const NavLink = styled(Link)`
-  color: #808080;
-  display: flex;
-  align-items: center;
-  text-decoration: none;
-  padding: 0 1rem;
-  height: 100%;
-  cursor: pointer;
-  &.active {
-    color: #000000;
-  }
-`
-
-// mobile menu icon styling
-const Bars = styled(FaBars)`
-  display: none;
-  color: #808080;
-
-  @media screen and (max-width: 768px) {
-    display: block;
-    position: absolute;
-    top: 0;
-    right: 0;
-    transform: translate(-100%, 75%);
-    font-size: 1.8rem;
-    cursor: pointer;
-  }
-`
-
-// container for navigation links
-const NavMenu = styled.div`
-  display: flex;
-  align-items: center;
-  margin-right: -24px;
-
-  /* Second Nav */
-  /* margin-right: 24px; */
-  /* Third Nav */
-  /* width: 100vw;
-  white-space: nowrap; */
-  @media screen and (max-width: 768px) {
-    display: none;
-  }
-`
-
-// container for navigation buttons
-const NavBtn = styled.nav`
-  display: flex;
-  align-items: center;
-  margin-right: 24px;
-  /* Third Nav */
-  /* justify-content: flex-end;
-  width: 100vw; */
-  @media screen and (max-width: 768px) {
-    display: none;
-  }
-`
-
-// styled component for navigation button links
-const NavBtnLink = styled(Link)`
-  border-radius: 4px;
-  background: #808080;
-  padding: 10px 22px;
-  color: #000000;
-  outline: none;
-  border: none;
-  cursor: pointer;
-  transition: all 0.2s ease-in-out;
-  text-decoration: none;
-  /* Second Nav */
-  margin-left: 24px;
-  &:hover {
-    transition: all 0.2s ease-in-out;
-    background: #fff;
-    color: #808080;
-  }
-`
-
-const Navbar = () => {
-  return (
-    <>
-      <Nav>
-        <Bars />
-
-        <NavMenu>
-          <NavLink to="/about">About</NavLink>
-          <NavLink to="/problems" activeStyle>
-            Problems
-          </NavLink>
-          <NavLink to="/sign-up" activeStyle>
-            Sign Up
-          </NavLink>
-          <NavLink to="/addProblems" activeStyle>
-            Add Problems
-          </NavLink>
-          {/* Second Nav */}
-          {/* <NavBtnLink to='/sign-in'>Sign In</NavBtnLink> */}
-        </NavMenu>
-        <NavBtn>
-          <NavBtnLink to="/signin">Sign In</NavBtnLink>
-        </NavBtn>
-      </Nav>
-    </>
-  )
-}
-
-export default Navbar
